Issue #130 - added "Magic Token" support to image names
- the [Manufacturer], [Year], etc. "Magic Tokens" for texts can
now all be used in the names of images as well.  The image will
update accordingly as the user navigates the frontend (for example,
fe.add_image("[Year]") will add an image that matches the year of
the current game (i.e. 1984.png, 1994.mp4, etc)
- Magic Tokens can now also be in the form [!<function_name>], in
which case Attract-Mode will run the specified function in squirrel
to get the string that the token is to be replaced with.
- See the updates to Layout.md for more about usage