Issue #151: shaders can now be specified as an absolute path
Shader paths should now behave the same as other files (image etc)
If an absolute path is provided, that is used.  If a relative path
is provided, it is assumed to be relative to the directory for the
layout/plugin/screensaver/intro that it is being loaded from.

Known issue: when specifying both a fragment and vertex shader, if
the first shader is specified as a relative path and is in a zip
file, the second shader is assumed to be in that zip file as well.