Issue #216 - separate UI menu navigation (up.down,left,right,back,sel)
- Added separate UI (up,down,left,right and back) controls that are configurable
in the controls configuration.
- A "default action" can be assigned to each of these in the controls config,
and by default it is the same as before: up is prev game, down is next game,
left is prev display, right is next display and back is exit (confirm).
- Added new "prev_game" and "next_game" actions, renamed "page_up" and
"page_down" to "prev_page" and "next_page" respectively.
- Related fixes to the filter, filter grid layouts and the submenu module
- UI controls must always be mapped.  If any one of them is not, it gets
reset to its default mapping.  The hope is that this will assist with
migrating from config files made with earlier versions of Attract-Mode.  This
will mess with some control configurations, which will have to be manually
corrected by the user.