Added list_size property to fe.ListBox
This property gets the size of the list shown by the listbox. When the listbox is assigned as an overlay custom control this property will return the number of options in the overlay dialog. It's useful if you want to adjust the overlay dialog's size, from the Transition.ShowOverlay, based on the number of the elements shown by the dialog.