Added kerning check to fit_string function
The function fit_string() now calculates the maximum string length in the same way as it's done in internal SFML Text::ensureGeometryUpdate() function.

We can now call text1.width = text1.msg_width and the width of the textbox will match the message width without cropping any characters.
This assumes nomargin is set to true and the new align modes are used.