True borderless window mode added
Borderless Fullscreen mode works in the same way as Window ( no border ) mode, but it sets the window size to the desktop resolution rather than previously set window size stored in window.am file. This way you get full optimization given by the system and no stuttering.
note: Needs to be checked on Linux and OSX, just to be sure it works as intended.