New missing image handler
Added a new option in settings that sets whether the default image placeholder is shown or a blank texture for specified image types. The default placeholder is embedded in a hpp file, but can be overridden by placing the default.png file inside the menu-art folder. The default setting is to not show the placeholder, but a blank texture instead.