Added Power Saving toggle
When it's on it works as before and helps to save power by only redrawing the screen when it's necessary. When off Attract keeps feeding new frames to the frame queue even if there was no change on the screen. It was reported that power saving set to on is causing animation stuttering due to stalling the frame queue and on Linux freezing of the entire application when videos are playing.