FPS Monitor Fix
Changed the way the graph is drawn due to some reported problems under DRM. FPS Monitor previously used a trick with surface clones to retrieve a previous frame content which allowed to make a scrolling buffer. Now it's drawn using the conventional method with image clones which is also as fast as before.