Added ScreenRefreshRate global variable
AM now is able to detect the refresh rate of the monitor. it takes just 200ms on AM start, but it does not utilize any platform specific functions. I think it's a good trade-off. This is useful for calculating the target speed of frame based animations on displays other than 60Hz. The advantage of frame based animations over time based is that they remain smooth when the frame delta time varies significantly on artwork reloading but there is no skipped frames due to triple buffering.

FPSMonitor plugin updates:
- not fixed to 60Hz anymore
- added frame time graph.
- fixed scaling issues when layout width, or height was overridden